Fixed issue with Joystick tester name.  Added CyController Example.
Joystick tester name had issues due to PSOC Creator is very picky about names of workspaces and their projects.  I had chaned the name of the project and workspace on the master branch, but the whole workspace was thrown out of whack.  I have now reloaded the project from its source and ensured it actually builds prior to committing.

CyController is a open-source example from a Cypress forums contributor.  This game controller was made for FIRST robotics, but is a general purpose HID controller.