Added Groundup Joystick, which is homemade by Prometheus Education
Added Joystick_Groundup.  This is a USB HID device with 2 axes of analog and 8 buttons, perfect for SNES.  The project can be expanded to include more buttons and axes if necessary.   Detailed instructions are included for making the USB project, minus the USB descriptor.  The USB descriptor can be found in the Joystick_Groundup project.